refactor: split amount.go into 4 files by concern
Split the 2,023-line amount.go into focused files:
- amount.go (650): types, constructors, accessors, JSON marshaling
- amount_arithmetic.go (607): Add, Sub, Compare, MulRatio, Mul, Div
- amount_round.go (540): MulRound, DivRound, and strict/native variants
- amount_sqrt.go (230): Sqrt, IOU division helpers

Pure file reorganization — zero behavioral changes.

Closes #78
